upgraded uninstall to erase everything and report success/failure properly

+++ bin/actions/uninstall_archive_action.js
@@ -6,9 +6,14 @@ const uninstallArchiveAction = (archive) => {
   return new Promise((resolve, reject) => {
     const uninstallPath = `${ApplicationConfiguration.get(`paths.archive`)}/${archive}`
 
-    FileSystem.removeDirectory(uninstallPath)
-    StatusLog.notify(`uninstalled`, archive)
-    resolve()
+    if (FileSystem.isDirectory(uninstallPath)) {
+      FileSystem.removeDirectory(uninstallPath)
+      StatusLog.notify(`uninstalled`, archive)
+      resolve()
+    }
+    else {
+      reject(`${uninstallPath} does not exist`)
+    }
   })
 }
